+++ lib/invar/private_file.rb
@@ -11,8 +11,13 @@ module Invar
       extend Forwardable
       def_delegators :@delegate_sd_obj, :stat, :to_s, :basename, :==, :chmod
 
+      # Mask for limiting to the lowest three octal digits (which store permissions)
+      PERMISSIONS_MASK = 0o777
+
       ALLOWED_USER_MODES  = [0o600, 0o400].freeze
       ALLOWED_GROUP_MODES = [0o060, 0o040, 0o000].freeze
+
+      DEFAULT_PERMISSIONS = 0o600
 
       # Allowed permissions modes for lockfile. Readable or read-writable by the user or group only
       ALLOWED_MODES = ALLOWED_USER_MODES.product(ALLOWED_GROUP_MODES).collect do |u, g|
@@ -45,9 +50,8 @@ module Invar
       #
       # @raise [FilePermissionsError] if the file has insecure permissions
       def verify_permissions!
-         permissions_mask = 0o777 # only the lowest three digits are perms, so masking
          # stat             = @delegate_sd_obj.stat
-         file_mode = stat.mode & permissions_mask
+         file_mode = stat.mode & PERMISSIONS_MASK
          # TODO: use stat.world_readable? etc instead
          return if ALLOWED_MODES.include? file_mode